heonil final

+++ lib/kernel/console.c
@@ -7,6 +7,7 @@
 #include "threads/interrupt.h"
 #include "threads/synch.h"
 
+
 static void vprintf_helper (char, void *);
 static void putchar_have_lock (uint8_t c);
 
@@ -140,7 +141,7 @@ puts (const char *s) {
 
 /* Writes the N characters in BUFFER to the console. */
 void
-putbuf (const char *buffer, size_t n) {
+putbuf (const char *buffer, long unsigned int n) {
 	acquire_console ();
 	while (n-- > 0)
 		putchar_have_lock (*buffer++);
